Refactor modules
The existing modules rely on the "pdnsapi" library which is proprietary
and has not received recent updates. Other Python libraries for
interfacing with the PowerDNS API suffer from the same issues or are
feature incomplete.
Solve this and make the modules more easily extensible by directly
interfacing with the PowerDNS HTTP REST API, utilizing the requests
library which is already installed as a dependency of a common Salt
deployment. For more efficient interaction as part of state modules
which repeatedly call the execution module, add the requests-toolbelt
library as the only additional dependency for easier implementation of
a base session.
As part of this refactoring the functions are renamed to follow a more
streamlined scheme differentiating between read and write operations,
and the output is aligned with what is returned directly by PowerDNS.
As a result, this is a breaking change.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>